moving source files around and changing GetToken to have a callback
